Set up proper pileline for Ubuntu
Set up github actions pipelines to generate docker images with basic
development tools.

This pipelines generates images for the following Ubuntu versions:
- 18.04
- 20.04

For each image, a number of tags are created, each pointing to
a version using specific compiler.

The list of compilers (and tags) is as folllows:
- gcc-10
- gcc-11
- clang-12
- clang-13

In addition a default image with the default compiler for that Ubuntu
version is also provided under the "default" tag

A set of development tools is also provided:
- cmake (latest)
- conan
- git

Finally, clang-tidy and clang format are included in all images except
the default ones.

The version installed depends on the compiler:
- clang-12 -> clang-format-12 / clang-tidy-12
- clang-13 -> clang-format-13 / clang-tidy-13
- gcc-10 -> clang-format-13 / clang-tidy-13
- gcc-11 -> clang-format-13 / clang-tidy-13